refactor: running eslint and fix most errors
