+++ package.json
@@ -6,6 +6,9 @@
   "author": "Elegant Themes",
   "license": "GPL2",
   "private": false,
+  "dependencies": {
+    "@wordpress/hooks": "^3.57.0"
+  },
   "devDependencies": {
     "@babel/core": "^7.24.5",
     "@babel/preset-env": "^7.24.5",
