+++ src/modal/registerDevClipboard.js
@@ -1,23 +1,22 @@
-import { DevClipboardContainer } from "./container";
-import {devClipboard} from '../icons';
-import { addFilter } from "@wordpress/hooks";
+import { addFilter } from '@wordpress/hooks';
+
+import { devClipboard } from '../icons';
+import { DevClipboardContainer } from './container';
+
 
 addFilter('divi.modalLibrary.modalMapping', 'divi', modals => {
   modals.DevClipboard = {
-    name: "divi/dev-clipboard",
-    label: "Clipboard",
-    type: "multiInstanceModal",
+    name:      'divi/dev-clipboard',
+    label:     'Clipboard',
+    type:      'multiInstanceModal',
     component: DevClipboardContainer,
   };
   return modals;
 });
 
 // Add module icons to the icon library.
-addFilter('divi.iconLibrary.icon.map', 'extensionExample', (icons) => {
-  return {
-    ...icons, // This is important. Without this, all other icons will be overwritten.
-    [devClipboard.name]:  devClipboard,
-    
-  };
-});
+addFilter('divi.iconLibrary.icon.map', 'extensionExample', icons => ({
+  ...icons, // This is important. Without this, all other icons will be overwritten.
+  [devClipboard.name]: devClipboard,
 
+}));
